Move viewset tests into own file
